feat: upgrade kotlin version to 1.4.0 & add coroutines library dependency & create LaunchApi and LaunchRepository
