+++ app/build.gradle
@@ -47,9 +47,13 @@ dependencies {
     // Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation "com.squareup.retrofit2:converter-scalars:2.9.0"
+    implementation 'com.squareup.retrofit2:converter-scalars:2.9.0'
 
     // OkHttp
-    implementation "com.squareup.okhttp3:okhttp:3.14.9"
-    implementation "com.squareup.okhttp3:logging-interceptor:3.14.9"
+    implementation 'com.squareup.okhttp3:okhttp:3.14.9'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.14.9'
+
+    // Coroutines
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.1'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.1"
 }
