+++ app/src/main/java/com/example/spaceapp/api/LaunchApi.kt
@@ -1,0 +1,11 @@
+package com.example.spaceapp.api
+
+import com.example.spaceapp.domain.entities.LaunchDTO
+import retrofit2.http.GET
+
+interface LaunchApi {
+
+    @GET("launch/upcoming")
+    suspend fun getUpcomingLaunches(): LaunchDTO
+
+}
