+++ app/src/main/java/com/example/spaceapp/domain/repositories/LaunchRepository.kt
@@ -1,0 +1,9 @@
+package com.example.spaceapp.domain.repositories
+
+import com.example.spaceapp.domain.entities.LaunchDTO
+
+interface LaunchRepository {
+
+    suspend fun getUpcomingLaunches(): LaunchDTO
+
+}
